kubernetes-bootstrap/requirements.txt
# Python dependencies for Kubernetes Bootstrap Scripts
PyYAML>=6.0
argparse
concurrent.futures